requirements.txt
pycryptodome==3.20.0
pyotp==2.9.0
qrcode==7.4.2
pillow==11.0.0
tkcalendar==1.6.1
pyzbar==0.1.8
cryptography==41.0.7